// src/sdks/computation/computationV2ApiDocs.ts
export interface EmissionStandard {
  companyId?: number;

  createBy?: number;

  createTime?: Date;

  dataList?: EmissionStandardData[];

  deleted?: boolean;

  endYear?: number;

  id?: number;

  orgId?: number;

  orgName?: string;

  policy?: string;

  settingType?: '1' | '2';

  standardEmission?: number;

  startYear?: number;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;
  languageSourceList?: any[];
}

export interface EmissionStandardData {
  createBy?: number;

  createTime?: Date;

  direct?: number;

  emissionStandardId?: number;

  energy?: number;

  id?: number;

  outsourcing?: number;

  rests?: number;

  scopeOne?: number;

  scopeThree?: number;

  scopeTwo?: number;

  supplyChain?: number;

  total?: number;

  transport?: number;

  updateBy?: number;

  updateTime?: Date;

  year?: number;
}